requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-stripe==2.10.3
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
packaging==25.0
pillow==12.0.0
pip-upgrade-tool==0.7.4
PyJWT==2.10.1
pyupgrade==3.21.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
stripe==13.0.1
tokenize_rt==6.2.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
